Add minion types and positions to combat
