+++ src/lib/types.ts
@@ -8,6 +8,7 @@ export * from './types/spell-types';
 export * from './types/equipment-types';
 export * from './types/wizard-types';
 export * from './types/enemy-types';
+export * from './types/minion-types';
 export * from './types/combat-types';
 export * from './types/game-types';
 export * from './types/market-types';
